Add files via upload
RDA5807 radio controlled by GD32VF103 chip. The code is in RISCV assembly. PA3,PA4,PA5,PA6 are used as inputs for tuneup , tunedown, seekup , seek down functions. The switches connected to these pins are press to close and connected to ground on one side. The function is called on detecting gnd on one of the pins. The display is ssd1306 128x32 OLED from aliexpress . The RDA5807 can be directly connceted to GD32vF103 I2C bus as both are 3.3v.  The OLED however has to be connected through some level convertrer as it is 5v. cheap options are available on aliexpress. PB6,PB7 are sck and sda respectievly. should be connceted with 4.7k to +3.3v. The above code was assembled with light weight BRONZEBEARD assembler written by Mr. Andrew. Can be downloaded from https://github.com/theandrew168/bronzebeard